add chuck norris as a plugin
